- fixed formatting error which occured for large files
git-svn-id: https://svn2.assembla.com/svn/savant/Savant/trunk@898 096565f4-0d05-4fcd-8cbf-ff6523290f6f

+++ src/savant/format/SavantFileFormatter.java
@@ -419,7 +419,7 @@ public class SavantFileFormatter {
 
     protected void writeReferenceMap(DataOutputStream f, List<String> refnames, Map<String, String> refnameToOutputFileMap) throws IOException {
 
-        int refOffset = 0;
+        long refOffset = 0;
 
         f.writeInt(refnames.size());
